test: fix MSVC /W4 warnings and Linux/Windows portability
Closes review item "we expect zero warnings from the test files"
under MSVC /W4 (the project enables it for MSVC builds in
CMakeLists.txt:36). All 18 test-file warnings fixed:

  C4100 (unreferenced parameter):
    - test_lifecycle.cpp: comment out names of unused callback
      lambda parameters

  C4996 (strdup deprecated):
    - test_common.h: add portable dasher_strdup() helper
      (_strdup on MSVC, strdup elsewhere)
    - test_multilingual.cpp, test_parameters.cpp: use it

  C4701 (uninitialized local):
    - test_capi.cpp: value-initialize dasher_parameter_info info{}
    - test_control_actions_navigation.cpp: initialize lb, hb to 0

  C4244 / C4267 (narrowing):
    - test_deterministic.cpp: cast strlen() to int
    - test_utf_conversion.cpp: explicit static_cast<int> on
      pointer arithmetic and strlen

Also fixed a real Linux/MSVC portability bug: test_view_geometry.cpp
used __INT_MAX__, a GCC/Clang internal macro not defined under MSVC.
Switched to standard INT_MAX from <climits>.

.gitignore: cover additional local scratch files generated during
this verification pass:
  - build-*/  (any alternate build dir, e.g. build-msvc/)
  - *.log.txt (build/run logs)
  - _*.ps1 / _*.bat / _*.sh / _*.py  (underscore-prefixed helpers)

Verified:
  - MSVC /W4 build of tests: 0 warnings (was 18)
  - gcc 14.2 build: clean, 0 warnings
  - clang-format clean

+++ tests/test_deterministic.cpp
@@ -66,7 +66,6 @@ TEST(det_different_y_different_output) {
     printf("  Y=100: '%s'\n", out2);
     // Different Y positions should generally produce different output
     ASSERT_NEQ(strcmp(out1, out2), 0);
-
 }
 
 TEST(det_no_input_no_output) {
@@ -136,7 +135,7 @@ TEST(det_speed_affects_rate) {
         dasher_set_screen_size(ctx, 800, 600);
         dasher_set_speed_percent(ctx, speeds[s]);
         const char* output = type_and_get(ctx, 700.0f, 300.0f, 100);
-        lens[s] = strlen(output);
+        lens[s] = (int)strlen(output);
         printf("  Speed %d%%: %d chars\n", speeds[s], lens[s]);
         dasher_destroy(ctx);
     }
@@ -144,5 +143,4 @@ TEST(det_speed_affects_rate) {
     // Higher speed should generally produce more text
     // (not strict for all cases, but 400% >> 100% should hold)
     ASSERT(lens[2] >= lens[0]);
-
 }
